% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCPRDConnection.R
\name{AbstractCPRDConnection}
\alias{AbstractCPRDConnection}
\title{Get a handle to the CPRD data databse}
\description{
Get a handle to the CPRD data databse

Get a handle to the CPRD data databse
}
\details{
This function sets up a connection to the cprd data databse, and provides a set of basic functions for manipulating the CPRD
}
\keyword{CPRD}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.con}}{- internal the mySQL connection}

\item{\code{.profile}}{- internal the connection details}

\item{\code{.dataDb}}{- internal the name of the cprd data database}

\item{\code{.analysisDb}}{- internal the name of the cprd analysis database}

\item{\code{.conventions}}{- internal the table naming conventions in the databases}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractCPRDConnection-new}{\code{AbstractCPRDConnection$new()}}
\item \href{#method-AbstractCPRDConnection-enableAdmin}{\code{AbstractCPRDConnection$enableAdmin()}}
\item \href{#method-AbstractCPRDConnection-namingConventions}{\code{AbstractCPRDConnection$namingConventions()}}
\item \href{#method-AbstractCPRDConnection-execSql}{\code{AbstractCPRDConnection$execSql()}}
\item \href{#method-AbstractCPRDConnection-fetchSql}{\code{AbstractCPRDConnection$fetchSql()}}
\item \href{#method-AbstractCPRDConnection-lazySql}{\code{AbstractCPRDConnection$lazySql()}}
\item \href{#method-AbstractCPRDConnection-appendDf}{\code{AbstractCPRDConnection$appendDf()}}
\item \href{#method-AbstractCPRDConnection-tableExists}{\code{AbstractCPRDConnection$tableExists()}}
\item \href{#method-AbstractCPRDConnection-conv}{\code{AbstractCPRDConnection$conv()}}
\item \href{#method-AbstractCPRDConnection-finalize}{\code{AbstractCPRDConnection$finalize()}}
\item \href{#method-AbstractCPRDConnection-print}{\code{AbstractCPRDConnection$print()}}
\item \href{#method-AbstractCPRDConnection-clone}{\code{AbstractCPRDConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-new}{}}}
\subsection{Method \code{new()}}{
Create a new CPRD data table connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$new(
  cprdEnv = getOption("cprd.environment", default = "test-analysis"),
  cprdConf = getOption("cprd.config", default = "~/.aurum.yaml"),
  cprdConnection = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cprdEnv}}{a profile for the connection.}

\item{\code{cprdConf}}{an absolute path to a .aurum.yaml configuration file.}

\item{\code{cprdConnection}}{an existing thing that inherits from AbstractCPRDConnection if one exists.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AbstractCPRDConnection object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-enableAdmin"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-enableAdmin}{}}}
\subsection{Method \code{enableAdmin()}}{
Internal function to switch to an admin role
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$enableAdmin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-namingConventions"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-namingConventions}{}}}
\subsection{Method \code{namingConventions()}}{
get the set of table names for this database and analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$namingConventions(analysisName = "test")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analysisName}}{the prefix for the analysis tables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-execSql"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-execSql}{}}}
\subsection{Method \code{execSql()}}{
Internal function to execute an SQL statement provided in the template substituting names as appropriate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$execSql(
  template,
  ...,
  analysisName = "test",
  varsList = self$namingConventions(analysisName),
  debug = getOption("cprd.debug", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template}}{an SQL glue template}

\item{\code{...}}{additional name=value for the glue substitution}

\item{\code{analysisName}}{the name of the analysis in progress if there is one}

\item{\code{varsList}}{(optional) all the names of the various configurable tables}

\item{\code{debug}}{log the generated SQL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-fetchSql"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-fetchSql}{}}}
\subsection{Method \code{fetchSql()}}{
Internal function to execute an SQL statement and fetch the query result provided in the template substituting names as appropriate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$fetchSql(
  template,
  ...,
  analysisName = "test",
  varsList = self$namingConventions(analysisName),
  debug = getOption("cprd.debug", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template}}{an SQL glue template}

\item{\code{...}}{additional name=value for the glue substitution}

\item{\code{analysisName}}{the name of the analysis in progress if there is one}

\item{\code{varsList}}{(optional) all the names of the various configurable tables}

\item{\code{debug}}{log the generated SQL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-lazySql"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-lazySql}{}}}
\subsection{Method \code{lazySql()}}{
Internal function to execute an SQL statement and wrap the query result in a lazy dataframe from the template substituting names as appropriate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$lazySql(
  template,
  ...,
  analysisName = "test",
  varsList = self$namingConventions(analysisName),
  debug = getOption("cprd.debug", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template}}{an SQL glue template}

\item{\code{...}}{additional name=value for the glue substitution}

\item{\code{analysisName}}{the name of the analysis in progress if there is one}

\item{\code{varsList}}{(optional) all the names of the various configurable tables}

\item{\code{debug}}{log the generated SQL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-appendDf"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-appendDf}{}}}
\subsection{Method \code{appendDf()}}{
Internal function to append a dataframe to a SQL table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$appendDf(df, table, database = self$.analysisDb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{a dataframe}

\item{\code{table}}{a table name}

\item{\code{database}}{the database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-tableExists"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-tableExists}{}}}
\subsection{Method \code{tableExists()}}{
Internal function to check if a table exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$tableExists(table, database = self$.analysisDb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{a table name}

\item{\code{database}}{the database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-conv"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-conv}{}}}
\subsection{Method \code{conv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$conv(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-finalize}{}}}
\subsection{Method \code{finalize()}}{
Shutdown the connection to the CPRD data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-print"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-print}{}}}
\subsection{Method \code{print()}}{
Prints the database connection information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCPRDConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCPRDConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCPRDConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
