% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{analysisSql}
\alias{analysisSql}
\title{config for the setup of of the CPRD database as installed by this package}
\format{
A list of lists:
}
\usage{
analysisSql
}
\description{
naming conventions and SQL configuration used during the setup of the CPRD database
this includes definitions of tables involved in an analysis.
}
\keyword{datasets}
