% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDBiomarkerAnalysis.R
\name{calculate_qdiabeteshf}
\alias{calculate_qdiabeteshf}
\title{calculate_qdiabeteshf}
\usage{
calculate_qdiabeteshf(
  dataframe,
  sex,
  age,
  ethrisk,
  town = NULL,
  smoking,
  duration,
  type1,
  cvd,
  af,
  renal,
  hba1c,
  cholhdl,
  sbp,
  bmi,
  surv = NULL
)
}
\arguments{
\item{sex}{- "male" or "female"}

\item{age}{- current age in years}

\item{ethrisk}{- QRISK2 ethnicity category: 0=Missing, 1=White, 2=Indian, 3=Pakistani, 4=Bangladeshi, 5=Other Asian, 6=Black Caribbean, 7=Black African, 8=Chinese, 9=Other ethnic group}

\item{town}{- Townsend Deprivation Index (default 0 i.e. missing)}

\item{smoking}{- QRISK2 smoking category: 0=Non-smoker, 1=Ex-smoker, 2=Current light smoker, 3=Current moderate smoker, 4=Current heavy smoker}

\item{duration}{- diabetes duration: 0=within the last year, 1=1-3 years, 2=4-6 years, 3=7-10 years, 4=11 or more}

\item{type1}{- Type 1 diabetes (binary) - otherwise Type 2 assumed}

\item{cvd}{- history of angina, heart attack or stroke}

\item{af}{- atrial fibrillation (binary)}

\item{renal}{- CKD stage 4 or 5}

\item{hba1c}{- last HbA1c in mmol/mol}

\item{cholhdl}{- cholesterol:HDL ratio}

\item{sbp}{- systolic blood pressure in mmHg}

\item{bmi}{- BMI in kg/m2}

\item{surv}{- how many years survival to use in model (default 10)}
}
\description{
calculates QDiabetes-HF 2015 score
}
