% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDBiomarkerAnalysis.R
\name{ckd_epi_2021_egfr}
\alias{ckd_epi_2021_egfr}
\title{Calculates eGFR from creatinine, sex and age at creatinine reading using 2021 CKD-EPI Creatinine (https://www.mdcalc.com/calc/3939/ckd-epi-equations-glomerular-filtration-rate-gfr#evidence)}
\usage{
ckd_epi_2021_egfr(dataset, creatinine_col, sex_col, age_at_creatinine_col)
}
\arguments{
\item{dataset}{- dataset containing creatinine, sex, and age at creatinine reading}

\item{creatinine_col}{- column with creatinine reading in umol/L}

\item{sex_col}{- column with sex: "male" or "female"}

\item{age_at_creatinine_col}{- column with age at creatinine reading in years}
}
\description{
see above
}
