% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDBiomarkerAnalysis.R
\name{clean_biomarker_units}
\alias{clean_biomarker_units}
\title{Clean biomarker units: only keep values with acceptable unit codes (includes missing unit code for all)}
\usage{
clean_biomarker_units(dataset, numunitid_col, biomrkr)
}
\arguments{
\item{dataset}{- dataset containing biomarker observations}

\item{biomrkr}{- name of biomarker to clean (acr/alt/ast/bmi/creatinine/dbp/fastingglucose/hba1c/hdl/height/ldl/pcr/sbp/smoking (for QRISK2)/totalcholesterol/triglyceride/weight)}
}
\description{
only keep measurements with 'acceptable' unit codes (numunitid) for specified biomarker
}
