% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDBiomarkerAnalysis.R
\name{clean_biomarker_values}
\alias{clean_biomarker_values}
\title{Clean biomarker values: only keep values within acceptable limits}
\usage{
clean_biomarker_values(dataset, biomrkr_col, biomrkr)
}
\arguments{
\item{dataset}{- dataset containing biomarker observations}

\item{biomrkr}{- name of biomarker to clean (acr/alt/ast/bmi/creatinine/dbp/fastingglucose/hba1c/hdl/height/ldl/pcr/sbp/totalcholesterol/triglyceride/weight)}
}
\description{
only keep measurements within acceptable limits for specified biomarker
}
