% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{codeSetsSql}
\alias{codeSetsSql}
\title{config for the setup of of the CPRD database as installed by this package}
\format{
A list of lists:
}
\usage{
codeSetsSql
}
\description{
naming conventions and SQL configuration used during the setup of the CPRD database
this includes definitions of the structure of codeSets.
}
\keyword{datasets}
