% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDData.R
\name{CPRDData}
\alias{CPRDData}
\title{Get a handle to the CPRD data databse}
\description{
Get a handle to the CPRD data databse

Get a handle to the CPRD data databse
}
\details{
This function sets up a connection to the cprd data databse, and provides a set of basic functions for manipulating the CPRD
}
\keyword{CPRD}
\section{Super class}{
\code{\link[aurum:AbstractCPRDConnection]{aurum::AbstractCPRDConnection}} -> \code{CPRDData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tables}}{- the raw database tables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CPRDData-new}{\code{CPRDData$new()}}
\item \href{#method-CPRDData-analysis}{\code{CPRDData$analysis()}}
\item \href{#method-CPRDData-codesets}{\code{CPRDData$codesets()}}
\item \href{#method-CPRDData-finalize}{\code{CPRDData$finalize()}}
\item \href{#method-CPRDData-print}{\code{CPRDData$print()}}
\item \href{#method-CPRDData-clone}{\code{CPRDData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="appendDf"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-appendDf'><code>aurum::AbstractCPRDConnection$appendDf()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="conv"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-conv'><code>aurum::AbstractCPRDConnection$conv()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="enableAdmin"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-enableAdmin'><code>aurum::AbstractCPRDConnection$enableAdmin()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="execSql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-execSql'><code>aurum::AbstractCPRDConnection$execSql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="fetchSql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-fetchSql'><code>aurum::AbstractCPRDConnection$fetchSql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="lazySql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-lazySql'><code>aurum::AbstractCPRDConnection$lazySql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="namingConventions"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-namingConventions'><code>aurum::AbstractCPRDConnection$namingConventions()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="tableExists"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-tableExists'><code>aurum::AbstractCPRDConnection$tableExists()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-new}{}}}
\subsection{Method \code{new()}}{
Create a new CPRD data table connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$new(
  cprdEnv = getOption("cprd.environment", default = "test-analysis"),
  cprdConf = getOption("cprd.config", default = "~/.aurum.yaml"),
  cprdConnection = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cprdEnv}}{a profile for the connection.}

\item{\code{cprdConf}}{an absolute path to a .aurum.yaml configuration file.}

\item{\code{cprdConnection}}{if a connection has already been established you can supply it here and it will be used instead. Anything than inherits from AbstractCPRDConnection will work. This is really for internal methods}

\item{\code{con}}{an existing mysql connection if one exists.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CPRDData object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-analysis"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-analysis}{}}}
\subsection{Method \code{analysis()}}{
Create or retrieve a cprd analysis container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$analysis(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the analysis - convention dictates this should be a lowercase name with underscores instead of spaces.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a cprd analysis container
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-codesets"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-codesets}{}}}
\subsection{Method \code{codesets()}}{
Retrieve the cprd codesets container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$codesets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a cprd codesets container
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-finalize}{}}}
\subsection{Method \code{finalize()}}{
Shutdown the connection to the CPRD data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-print"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-print}{}}}
\subsection{Method \code{print()}}{
Prints the database connection information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
