% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{dataSql}
\alias{dataSql}
\title{config for the setup of of the CPRD database as installed by this package}
\format{A list of lists:}
\usage{
dataSql
}
\description{
naming conventions and SQL configuration used during the setup of the CPRD database
this includes raw data tables from the CPRD.
}
\keyword{datasets}
