% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{lookupSql}
\alias{lookupSql}
\title{config for the setup of of the CPRD database as installed by this package}
\format{A list of lists:}
\usage{
lookupSql
}
\description{
naming conventions and SQL configuration used during the setup of the CPRD database
this includes reference data tables such as code lists.
}
\keyword{datasets}
