% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{operationalSql}
\alias{operationalSql}
\title{config for the setup of of the CPRD database as installed by this package}
\format{
A list of lists:
}
\usage{
operationalSql
}
\description{
naming conventions and SQL configuration used during the setup of the CPRD database
this includes operational tables such as log tables.
}
\keyword{datasets}
