% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDCodeSets.R
\name{CPRDCodeSets}
\alias{CPRDCodeSets}
\title{Get a handle to the CPRD data databse}
\description{
Get a handle to the CPRD data databse

Get a handle to the CPRD data databse
}
\details{
This function sets up a connection to the cprd data databse, and provides a set of basic functions for manipulating the CPRD
}
\keyword{CPRD}
\section{Super class}{
\code{\link[aurum:AbstractCPRDConnection]{aurum::AbstractCPRDConnection}} -> \code{CPRDCodeSets}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.data}}{- internal the pointer to the cprdDatabase}

\item{\code{codeSets}}{- the med code sets table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CPRDCodeSets-new}{\code{CPRDCodeSets$new()}}
\item \href{#method-CPRDCodeSets-loadMedCodeSet}{\code{CPRDCodeSets$loadMedCodeSet()}}
\item \href{#method-CPRDCodeSets-loadProdCodeSet}{\code{CPRDCodeSets$loadProdCodeSet()}}
\item \href{#method-CPRDCodeSets-loadICD10CodeSet}{\code{CPRDCodeSets$loadICD10CodeSet()}}
\item \href{#method-CPRDCodeSets-loadOPCS4CodeSet}{\code{CPRDCodeSets$loadOPCS4CodeSet()}}
\item \href{#method-CPRDCodeSets-loadCodeSet}{\code{CPRDCodeSets$loadCodeSet()}}
\item \href{#method-CPRDCodeSets-loadAll}{\code{CPRDCodeSets$loadAll()}}
\item \href{#method-CPRDCodeSets-getCodeSetDetails}{\code{CPRDCodeSets$getCodeSetDetails()}}
\item \href{#method-CPRDCodeSets-getCodeSet}{\code{CPRDCodeSets$getCodeSet()}}
\item \href{#method-CPRDCodeSets-getAllCodeSetVersion}{\code{CPRDCodeSets$getAllCodeSetVersion()}}
\item \href{#method-CPRDCodeSets-listCodeSets}{\code{CPRDCodeSets$listCodeSets()}}
\item \href{#method-CPRDCodeSets-deleteCodeSetVersion}{\code{CPRDCodeSets$deleteCodeSetVersion()}}
\item \href{#method-CPRDCodeSets-deleteCodeSet}{\code{CPRDCodeSets$deleteCodeSet()}}
\item \href{#method-CPRDCodeSets-print}{\code{CPRDCodeSets$print()}}
\item \href{#method-CPRDCodeSets-clone}{\code{CPRDCodeSets$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="appendDf"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-appendDf'><code>aurum::AbstractCPRDConnection$appendDf()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="conv"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-conv'><code>aurum::AbstractCPRDConnection$conv()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="enableAdmin"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-enableAdmin'><code>aurum::AbstractCPRDConnection$enableAdmin()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="execSql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-execSql'><code>aurum::AbstractCPRDConnection$execSql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="fetchSql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-fetchSql'><code>aurum::AbstractCPRDConnection$fetchSql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="lazySql"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-lazySql'><code>aurum::AbstractCPRDConnection$lazySql()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="namingConventions"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-namingConventions'><code>aurum::AbstractCPRDConnection$namingConventions()</code></a></li>
<li><span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="tableExists"><a href='../../aurum/html/AbstractCPRDConnection.html#method-AbstractCPRDConnection-tableExists'><code>aurum::AbstractCPRDConnection$tableExists()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-new"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-new}{}}}
\subsection{Method \code{new()}}{
Get a new CPRD codes sets container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$new(cprdData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cprdData}}{an existing cprd data container}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CPRDCodeSets container.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadMedCodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadMedCodeSet}{}}}
\subsection{Method \code{loadMedCodeSet()}}{
load a new medcode set into the CPRD data from an R dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadMedCodeSet(
  codeSetDf,
  category = NULL,
  name,
  version,
  colname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codeSetDf}}{a dataframe with minimally a single column of medcode ids.}

\item{\code{category}}{(optional) column containing category code. If set to NULL, all values are NULL.}

\item{\code{name}}{the name of the codeset}

\item{\code{version}}{the version of the codeset}

\item{\code{colname}}{(optional) the name of the column containing the medcode. If set to NULL defaults to the first column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TRUE/FALSE depending on if the code set was successfully loaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadProdCodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadProdCodeSet}{}}}
\subsection{Method \code{loadProdCodeSet()}}{
load a new prodcode set into the CPRD data from an R dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadProdCodeSet(
  codeSetDf,
  category = NULL,
  name,
  version,
  colname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codeSetDf}}{a dataframe with minimally a single column of prodcode ids.}

\item{\code{category}}{(optional) column containing category code. If set to NULL, all values are NULL.}

\item{\code{name}}{the name of the codeset}

\item{\code{version}}{the version of the codeset}

\item{\code{colname}}{(optional) the name of the column containing the prodcode. If set to NULL defaults to the first column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TRUE/FALSE depending on if the code set was successfully loaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadICD10CodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadICD10CodeSet}{}}}
\subsection{Method \code{loadICD10CodeSet()}}{
load a new ICD10 set into the CPRD data from an R dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadICD10CodeSet(
  codeSetDf,
  category = NULL,
  name,
  version,
  colname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codeSetDf}}{a dataframe with minimally a single column of ICD10 codes.}

\item{\code{category}}{(optional) column containing category code. If set to NULL, all values are NULL.}

\item{\code{name}}{the name of the codeset}

\item{\code{version}}{the version of the codeset}

\item{\code{colname}}{(optional) the name of the column containing the ICD10 code If set to NULL defaults to the first column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TRUE/FALSE depending on if the code set was successfully loaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadOPCS4CodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadOPCS4CodeSet}{}}}
\subsection{Method \code{loadOPCS4CodeSet()}}{
load a new ICD10 set into the CPRD data from an R dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadOPCS4CodeSet(
  codeSetDf,
  category = NULL,
  name,
  version,
  colname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codeSetDf}}{a dataframe with minimally a single column of ICD10 codes.}

\item{\code{category}}{(optional) column containing category code. If set to NULL, all values are NULL.}

\item{\code{name}}{the name of the codeset}

\item{\code{version}}{the version of the codeset}

\item{\code{colname}}{(optional) the name of the column containing the ICD10 code If set to NULL defaults to the first column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TRUE/FALSE depending on if the code set was successfully loaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadCodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadCodeSet}{}}}
\subsection{Method \code{loadCodeSet()}}{
load a new code set into the CPRD data from an R dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadCodeSet(
  codeSetDf,
  category = NULL,
  name,
  version,
  colname = NULL,
  type
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codeSetDf}}{a dataframe with minimally a single column of medcode ids or prodcode ids.}

\item{\code{category}}{(optional) column containing category code. If set to NULL defaults to the third column.}

\item{\code{name}}{the name of the codeset}

\item{\code{version}}{the version of the codeset}

\item{\code{colname}}{the name of the column containing the medcode or prodcode. If set to NULL defaults to the first column.}

\item{\code{type}}{the type of the code as the column name it will match on e.g. medcodeid, prodcodeid. This can be used for any kind of codeid column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TRUE/FALSE depending on if the code set was successfully loaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-loadAll"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-loadAll}{}}}
\subsection{Method \code{loadAll()}}{
load a set of med codes from tab separated text files into the CPRD database.
The names of the codesets will be defined by the filenames. Files will not be loaded if they have the same hash,
name and version as an existing code set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$loadAll(paths, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paths}}{a list of file paths to try and load.}

\item{\code{version}}{the version of the codesets that are being loaded}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-getCodeSetDetails"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-getCodeSetDetails}{}}}
\subsection{Method \code{getCodeSetDetails()}}{
list the codes, and their descriptions, in a given named version of a code set. The medcodeid or prodcodeid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$getCodeSetDetails(name, category = NULL, version = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the code set name}

\item{\code{version}}{the code set version}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a lazy dataframe containing the codes and their descriptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-getCodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-getCodeSet}{}}}
\subsection{Method \code{getCodeSet()}}{
list the codes only in a named code set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$getCodeSet(name, category = NULL, version = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the code set name}

\item{\code{version}}{(optional) the code set version, if not sepcified returns the maximium code}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a lazy dataframe containing the code only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-getAllCodeSetVersion"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-getAllCodeSetVersion}{}}}
\subsection{Method \code{getAllCodeSetVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$getAllCodeSetVersion(version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-listCodeSets"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-listCodeSets}{}}}
\subsection{Method \code{listCodeSets()}}{
list the codes only in a named code set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$listCodeSets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a lazy dataframe containing the code only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-deleteCodeSetVersion"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-deleteCodeSetVersion}{}}}
\subsection{Method \code{deleteCodeSetVersion()}}{
delete a code set version. This will affect all users of the CPRD database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$deleteCodeSetVersion(name, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the code set name}

\item{\code{version}}{the code set version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-deleteCodeSet"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-deleteCodeSet}{}}}
\subsection{Method \code{deleteCodeSet()}}{
delete all versions of a code set. This will affect all users of the CPRD database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$deleteCodeSet(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the code set name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-print"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-print}{}}}
\subsection{Method \code{print()}}{
Prints the database connection information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CPRDCodeSets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CPRDCodeSets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDCodeSets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
