% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDData.R
\name{CPRDData}
\alias{CPRDData}
\title{Get a handle to the CPRD data databse}
\description{
Get a handle to the CPRD data databse

Get a handle to the CPRD data databse
}
\details{
This function sets up a connection to the cprd data databse, and provides a set of basic functions for manipulating the CPRD
}
\keyword{CPRD}
\section{Super class}{
\code{\link[aurum:AbstractCPRDConnection]{aurum::AbstractCPRDConnection}} -> \code{CPRDData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tables}}{- the raw database tables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CPRDData$new()}}
\item \href{#method-analysis}{\code{CPRDData$analysis()}}
\item \href{#method-codesets}{\code{CPRDData$codesets()}}
\item \href{#method-finalize}{\code{CPRDData$finalize()}}
\item \href{#method-print}{\code{CPRDData$print()}}
\item \href{#method-clone}{\code{CPRDData$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="appendDf">}\href{../../aurum/html/AbstractCPRDConnection.html#method-appendDf}{\code{aurum::AbstractCPRDConnection$appendDf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="conv">}\href{../../aurum/html/AbstractCPRDConnection.html#method-conv}{\code{aurum::AbstractCPRDConnection$conv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="enableAdmin">}\href{../../aurum/html/AbstractCPRDConnection.html#method-enableAdmin}{\code{aurum::AbstractCPRDConnection$enableAdmin()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="execSql">}\href{../../aurum/html/AbstractCPRDConnection.html#method-execSql}{\code{aurum::AbstractCPRDConnection$execSql()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="fetchSql">}\href{../../aurum/html/AbstractCPRDConnection.html#method-fetchSql}{\code{aurum::AbstractCPRDConnection$fetchSql()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="lazySql">}\href{../../aurum/html/AbstractCPRDConnection.html#method-lazySql}{\code{aurum::AbstractCPRDConnection$lazySql()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="namingConventions">}\href{../../aurum/html/AbstractCPRDConnection.html#method-namingConventions}{\code{aurum::AbstractCPRDConnection$namingConventions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="aurum" data-topic="AbstractCPRDConnection" data-id="tableExists">}\href{../../aurum/html/AbstractCPRDConnection.html#method-tableExists}{\code{aurum::AbstractCPRDConnection$tableExists()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new CPRD data table connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$new(
  cprdEnv = getOption("cprd.environment", default = "test-analysis"),
  cprdConf = getOption("cprd.config", default = "~/.aurum.yaml"),
  cprdConnection = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cprdEnv}}{a profile for the connection.}

\item{\code{cprdConf}}{an absolute path to a .aurum.yaml configuration file.}

\item{\code{cprdConnection}}{if a connection has already been established you can supply it here and it will be used instead. Anything than inherits from AbstractCPRDConnection will work. This is really for internal methods}

\item{\code{con}}{an existing mysql connection if one exists.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CPRDData object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-analysis"></a>}}
\if{latex}{\out{\hypertarget{method-analysis}{}}}
\subsection{Method \code{analysis()}}{
Create or retrieve a cprd analysis container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$analysis(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the analysis - convention dictates this should be a lowercase name with underscores instead of spaces.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a cprd analysis container
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codesets"></a>}}
\if{latex}{\out{\hypertarget{method-codesets}{}}}
\subsection{Method \code{codesets()}}{
Retrieve the cprd codesets container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$codesets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a cprd codesets container
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Shutdown the connection to the CPRD data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints the database connection information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPRDData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
